Backend/requirements.txt
Django==5.1.4
pymongo==4.3.3
mongoengine==0.29.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt-5.3.1 
pyjwt-2.10.1
python-dotenv
jwt